flask_file_upload/requirements.txt
Flask==3.0.3
requests==2.32.3
gunicorn==21.2.0  # For serving Flask in production
Flask-CORS==3.0.10
python-dotenv==1.0.1